Tag search works on new ui
